Update shards to latest versions, and disable onyx-sql & Avram
This combination compiles using Crystal 1.12.2.

Onyx appears to be unmaintained and needs updates to work with the latest Crystal.

Requiring avram breaks compilation of granite test, I suspect due to the monkey patching of `JSON::Serializable` in `avram/charms/json_extensions.cr`. It's therefore possible to run one or the other, but not both.